+++ src/main/java/com/example/crmenercom/entity/OrderEntity.java
@@ -15,9 +15,9 @@ import java.util.List;
 public class OrderEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private LocalDate date;
-    private int customerId;
+    private Long customerId;
     private Integer status;
 
     @ManyToMany
