+++ src/main/java/com/example/crmenercom/repository/ContactPersonRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactPersonRepository extends JpaRepository<ContactPersonEntity, Integer> {
+public interface ContactPersonRepository extends JpaRepository<ContactPersonEntity, Long> {
 }
