+++ src/main/java/com/example/crmenercom/repository/ContactedByRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactedByRepository extends JpaRepository<ContactedByEntity, Integer> {
+public interface ContactedByRepository extends JpaRepository<ContactedByEntity, Long> {
 }
