+++ src/main/java/com/example/crmenercom/repository/SalesQuantityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalesQuantityRepository extends JpaRepository<SalesQuantityEntity, Integer> {
+public interface SalesQuantityRepository extends JpaRepository<SalesQuantityEntity, Long> {
 }
